pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-ui"
version = "0.1.0"
description = "Research desk backend"
requires-python = ">=3.10"
dependencies = [
  "duckdb>=1.0.0",
  "pyyaml>=6.0.1",
  "starlette>=0.37.2",
  "uvicorn>=0.30.0",
]

[tool.setuptools]
packages = ["app"]
